Reimplement save/load functionality
- The previous game save database has been rewritten using Room. The old
database is not migrated, so existing saves may not longer be discoverable.
That's not a huge concern though, as they're so out of date that they won't
be usable with the new CTH engine anyway, and there's no plans to support
them.
- Save games can be deleted by swiping left on either the save or load
views. There is a grace period during which the deletion can be undone.
- Add screenshots and save game locations to the configuration. These are
both stored in the app's file storage, and may be backed up.
- Autosave games appear when loading a game but do not appear when saving,
to prevent them from being overwritten.
- A confirmation is presented when attempting to overwrite an existing
save game.